newdesignbabyclare/requirements.txt
opencv-python==4.5.5.64
numpy==1.22.0
sounddevice==0.4.4
wavio==0.0.4
torch==1.10.1
torchaudio==0.10.1
scipy==1.7.3
requests>=2.28.2
hume==0.2.0
pydub==0.25.1
python-dotenv==0.19.2
pytest==6.2.5
pyyaml==6.0.1
toml==0.10.2
iniconfig==2.0.0
pluggy<2.0,>=0.12
py>=1.8.2
keras==2.12.0
keras-applications==1.0.8
jax==0.3.15
libclang==13.0.0
git+https://github.com/rcmalli/keras-vggface.git

# Additional setup instructions for ffmpeg (required for pydub)
# 
# macOS:
# You can install ffmpeg using Homebrew:
# brew install ffmpeg
#
# Ubuntu:
# You can install ffmpeg using apt:
# sudo apt update
# sudo apt install ffmpeg
#
# Windows:
# You can download ffmpeg from the official FFmpeg website (https://ffmpeg.org/download.html)
# and add it to your system's PATH.
